Issue 846: Switch from Texen to Groovy for DB script generation. (#872)

+++ app/src/main/java/org/tightblog/business/search/tasks/SearchTask.java
@@ -145,7 +145,7 @@ public class SearchTask extends AbstractTask {
         if (searchResults == null) {
             return -1;
         }
-        return searchResults.totalHits;
+        return (int) searchResults.totalHits;
     }
 
     public String getTerm() {
